requirements.txt
InstructorEmbedding==1.0.1
langchain==0.3.9
langchain_community==0.3.9
PyPDF2==3.0.1
python-dotenv==1.0.1
sentence_transformers==3.2.1
streamlit==1.39.0
openai
tiktoken
faiss-cpu